+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - DOM tool: Fix z-index stacking issue
 - Layout: Add option to hide the battle mode pill
+- Core: Fix the need to reload page to see tools
 
 # 1.2.4